refact: k8s dashboard replaced with headlamp

+++ tofu/envs/dev/destroy.sh
@@ -8,4 +8,11 @@ set -o allexport
 set +o allexport
 . ./secrets/openrc
 
+read -p "WARNING! Current deployment will be destroyed. Are you sure you want to proceed, y/n? " -n 1 -r
+echo
+if [[ ! $REPLY =~ ^[Yy]$ ]]
+then
+    [[ "$0" = "$BASH_SOURCE" ]] && exit 1 || return 1
+fi
+
 tofu destroy
